tail.Tailer uses own errgroup instance per source
Move tail.Tailer to its own source file and have Tailer own its own
errgroup. This prevents errors from one tailing process from impacting
other tails and their own errgroups.
