+++ collector/logs/sources/tail/tail.go
@@ -4,20 +4,15 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/Azure/adx-mon/collector/logs/engine"
 	"github.com/Azure/adx-mon/collector/logs/sources/tail/sourceparse"
-	"github.com/Azure/adx-mon/collector/logs/transforms/parser"
 	"github.com/Azure/adx-mon/collector/logs/types"
 	"github.com/Azure/adx-mon/pkg/logger"
-	"github.com/tenebris-tech/tail"
-	"golang.org/x/sync/errgroup"
 )
 
 const (
@@ -31,7 +26,7 @@ var (
 )
 
 // FileTailTarget describes a file to tail, how to parse it, and the destination for the parsed logs.
-// It is used int he list of StaticTargets in TailSourceConfig, or used in AddTarget.
+// It is used in the list of StaticTargets in TailSourceConfig, or used in AddTarget.
 // NOTE: If new fields are added that change during runtime (e.g. from pod metadata changing) that are sent
 // via UpdateChan, update isTargetChanged to take this new field into account.
 type FileTailTarget struct {
@@ -69,23 +64,6 @@ type TailSourceConfig struct {
 	PodDiscoveryOpts *PodDiscoveryOpts
 }
 
-// Tailer is a specific instance of a file being tailed.
-type Tailer struct {
-	tail           *tail.Tail
-	shutdown       context.CancelFunc
-	database       string
-	table          string
-	logTypeParser  sourceparse.LogTypeParser
-	logLineParsers []parser.Parser
-	resources      map[string]interface{}
-}
-
-func (t *Tailer) Stop() {
-	t.tail.Cleanup()
-	t.tail.Stop()
-	t.shutdown()
-}
-
 // TailSource implements the types.Source interface for tailing files.
 type TailSource struct {
 	staticTargets   []FileTailTarget
@@ -94,8 +72,6 @@ type TailSource struct {
 
 	mu           sync.RWMutex
 	closeFn      context.CancelFunc
-	groupCtx     context.Context
-	group        *errgroup.Group
 	ackGenerator func(*types.Log) func()
 	tailers      map[string]*Tailer
 
@@ -120,10 +96,6 @@ func (s *TailSource) Open(ctx context.Context) error {
 	ctx, closeFn := context.WithCancel(ctx)
 	s.closeFn = closeFn
 
-	group, ctx := errgroup.WithContext(ctx)
-	s.group = group
-	s.groupCtx = ctx
-
 	s.ackGenerator = noopAckGenerator
 	if s.cursorDirectory != "" {
 		s.ackGenerator = func(log *types.Log) func() {
@@ -145,6 +117,7 @@ func (s *TailSource) Open(ctx context.Context) error {
 			// On startup, if we fail to add a target, we should close all the tailers we've opened so far and return.
 			for _, t := range s.tailers {
 				t.Stop()
+				t.Wait()
 			}
 			return fmt.Errorf("TailSource open: %w", err)
 		}
@@ -156,6 +129,7 @@ func (s *TailSource) Open(ctx context.Context) error {
 			// On startup, if we fail to open the pod discovery, we should close all the tailers we've opened so far and return.
 			for _, t := range s.tailers {
 				t.Stop()
+				t.Wait()
 			}
 			return fmt.Errorf("TailSource open: %w", err)
 		}
@@ -173,11 +147,13 @@ func (s *TailSource) Close() error {
 	for _, t := range s.tailers {
 		t.Stop()
 	}
+	for _, t := range s.tailers {
+		t.Wait()
+	}
 	clear(s.tailers)
 	s.mu.Unlock()
 
 	s.closeFn()
-	s.group.Wait()
 	return nil
 }
 
@@ -196,66 +172,19 @@ func (s *TailSource) AddTarget(target FileTailTarget, updateChan <-chan FileTail
 		return nil // already exists
 	}
 
-	tailerCtx, shutdown := context.WithCancel(s.groupCtx)
-	batchQueue := make(chan *types.Log, 512)
-	outputQueue := make(chan *types.LogBatch, 1)
-	tailConfig := tail.Config{Follow: true, ReOpen: true, Poll: true}
-	existingCursorPath := cursorPath(s.cursorDirectory, target.FilePath)
-	fileId, position, err := readCursor(existingCursorPath)
-	if err == nil {
-		if logger.IsDebug() {
-			logger.Debugf("TailSource: found existing cursor for file %q: %s %d", target.FilePath, fileId, position)
-		}
-		tailConfig.Location = &tail.SeekInfo{
-			Offset:         position,
-			Whence:         io.SeekStart,
-			FileIdentifier: fileId,
-		}
+	tailerConfig := TailerConfig{
+		Target:          target,
+		UpdateChan:      updateChan,
+		AckGenerator:    s.ackGenerator,
+		WorkerCreator:   s.workerCreator,
+		CursorDirectory: s.cursorDirectory,
+		WorkerName:      s.Name(),
 	}
 
-	tailFile, err := tail.TailFile(target.FilePath, tailConfig)
+	tailer, err := StartTailing(tailerConfig)
 	if err != nil {
-		shutdown()
-		return fmt.Errorf("addTarget create tailfile: %w", err)
+		return fmt.Errorf("AddTarget: %w", err)
 	}
-
-	parsers, err := parser.NewParsers(target.Parsers)
-	if err != nil {
-		shutdown()
-		return fmt.Errorf("addTarget create parsers: %w", err)
-	}
-
-	attributes := make(map[string]interface{})
-	for k, v := range target.Resources {
-		attributes[k] = v
-	}
-
-	tailer := &Tailer{
-		tail:           tailFile,
-		shutdown:       shutdown,
-		database:       target.Database,
-		table:          target.Table,
-		logTypeParser:  sourceparse.GetLogTypeParser(target.LogType),
-		logLineParsers: parsers,
-		resources:      attributes,
-	}
-	s.group.Go(func() error {
-		return readLines(tailerCtx, tailer, updateChan, batchQueue)
-	})
-
-	batchConfig := engine.BatchConfig{
-		MaxBatchSize: 1000,
-		MaxBatchWait: 1 * time.Second,
-		InputQueue:   batchQueue,
-		OutputQueue:  outputQueue,
-		AckGenerator: s.ackGenerator,
-	}
-	s.group.Go(func() error {
-		return engine.BatchLogs(tailerCtx, batchConfig)
-	})
-
-	worker := s.workerCreator(s.Name(), outputQueue)
-	s.group.Go(worker.Run)
 
 	s.tailers[target.FilePath] = tailer
 
@@ -271,90 +200,9 @@ func (s *TailSource) RemoveTarget(filePath string) {
 	tailer, ok := s.tailers[filePath]
 	if ok {
 		tailer.Stop()
+		tailer.Wait()
 		delete(s.tailers, filePath)
 		cleanCursor(cursorPath(s.cursorDirectory, filePath))
-	}
-}
-
-func readLines(ctx context.Context, tailer *Tailer, updateChannel <-chan FileTailTarget, outputQueue chan<- *types.Log) error {
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		// Receive updates from the optional updateChannel.
-		case newTarget, ok := <-updateChannel:
-			if ok {
-				newParsers, err := parser.NewParsers(newTarget.Parsers)
-				if err != nil {
-					logger.Errorf("readLines: parser error for filename %q: %v", tailer.tail.Filename, err)
-					continue
-				}
-				tailer.logLineParsers = newParsers
-				tailer.database = newTarget.Database
-				tailer.table = newTarget.Table
-				tailer.resources = make(map[string]interface{})
-				for k, v := range newTarget.Resources {
-					tailer.resources[k] = v
-				}
-			}
-		case line, ok := <-tailer.tail.Lines:
-			if !ok {
-				logger.Infof("readLines: tailer closed the channel for filename %q", tailer.tail.Filename)
-				return nil // No longer getting lines due to the tailer being closed. Exit.
-			}
-			if line.Err != nil {
-				logger.Errorf("readLines: tailer error for filename %q: %v", tailer.tail.Filename, line.Err)
-				//skip
-				continue
-			}
-
-			log := types.LogPool.Get(1).(*types.Log)
-			log.Reset()
-
-			isPartial, err := tailer.logTypeParser.Parse(line.Text, log)
-			if err != nil {
-				logger.Errorf("readLines: parselog error for filename %q: %v", tailer.tail.Filename, err)
-				//skip
-				types.LogPool.Put(log)
-				continue
-			}
-			if isPartial {
-				types.LogPool.Put(log)
-				continue
-			}
-
-			position := line.Offset
-			currentFileId := line.FileIdentifier
-			log.Attributes[types.AttributeDatabaseName] = tailer.database
-			log.Attributes[types.AttributeTableName] = tailer.table
-
-			for k, v := range tailer.resources {
-				log.Resource[k] = v
-			}
-
-			successfulParse := false
-			for _, parser := range tailer.logLineParsers {
-				err := parser.Parse(log)
-				if err == nil {
-					successfulParse = true
-					break // successful parse
-				} else if logger.IsDebug() {
-					logger.Debugf("readLines: parser error for filename %q: %v", tailer.tail.Filename, err)
-				}
-			}
-
-			if successfulParse {
-				// Successful parse, remove the raw message
-				delete(log.Body, types.BodyKeyMessage)
-			}
-
-			// Write after parsing to ensure these values are always set to values we need for acking.
-			log.Attributes[cursor_position] = position
-			log.Attributes[cursor_file_id] = currentFileId
-			log.Attributes[cursor_file_name] = tailer.tail.Filename
-
-			outputQueue <- log
-		}
 	}
 }
